chapter4/langserve_sample/requirements.txt
langchain==0.3.7
langchain-openai==0.2.9
langchain-community==0.3.7
langchain-core==0.3.18
langchain-text-splitters==0.3.2
langchain-pinecone==0.2.0
langserve[all]==0.3.0
httpx==0.27.0
pypdf==5.0.1
pinecone==5.3.1
pycryptodome==3.21.0
python-dotenv==1.0.1